+++ .env
@@ -1,0 +1,4 @@
+API_NAME=sga_espacio_academico_mid
+SERVICE_NAME=sgaespacioacademicomid
+API_BASE_DIR=github.com/udistrital
+SGA_ESPACIO_ACADEMICO_MID_HTTP_PORT=8095
