+++ conf/app.conf
@@ -1,5 +1,5 @@
 appname = sga_espacio_academico_mid
-httpport = ${SGA__ESPACIOS_ACADEMICOS_MID_HTTP_PORT}
+httpport = ${SGA_ESPACIO_ACADEMICO_MID_HTTP_PORT}
 runmode = dev
 autorender = false
 copyrequestbody = true
